extracted ContainerBuilderBuilder
