+++ src/Kernel/ContainerBuilderBuilder.php
@@ -1,0 +1,71 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Core\Kernel;
+
+use Rector\Core\Config\Loader\ConfigureCallMergingLoaderFactory;
+use Rector\Core\DependencyInjection\Collector\ConfigureCallValuesCollector;
+use Rector\Core\DependencyInjection\CompilerPass\AutowireArrayParameterCompilerPass;
+use Rector\Core\DependencyInjection\CompilerPass\AutowireRectorCompilerPass;
+use Rector\Core\DependencyInjection\CompilerPass\MakeRectorsPublicCompilerPass;
+use Rector\Core\DependencyInjection\CompilerPass\MergeImportedRectorConfigureCallValuesCompilerPass;
+use Rector\Core\DependencyInjection\CompilerPass\RemoveSkippedRectorsCompilerPass;
+use Rector\Core\Util\FileHasher;
+use Rector\Testing\PHPUnit\StaticPHPUnitEnvironment;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Webmozart\Assert\Assert;
+
+final class ContainerBuilderBuilder {
+
+    private readonly ConfigureCallValuesCollector $configureCallValuesCollector;
+
+    public function __construct()
+    {
+        $this->configureCallValuesCollector = new ConfigureCallValuesCollector();
+    }
+
+    /**
+     * @param string[] $configFiles
+     */
+    public function build(array $configFiles): \Symfony\Component\DependencyInjection\ContainerBuilder
+    {
+        $compilerPasses = $this->createCompilerPasses();
+
+        $configureCallMergingLoaderFactory = new ConfigureCallMergingLoaderFactory($this->configureCallValuesCollector);
+        $containerBuilderFactory = new ContainerBuilderFactory($configureCallMergingLoaderFactory);
+
+        $containerBuilder = $containerBuilderFactory->create($configFiles, $compilerPasses);
+
+        // @see https://symfony.com/blog/new-in-symfony-4-4-dependency-injection-improvements-part-1
+        $containerBuilder->setParameter('container.dumper.inline_factories', true);
+        // to fix reincluding files again
+        $containerBuilder->setParameter('container.dumper.inline_class_loader', false);
+
+        $containerBuilder->compile();
+
+        return $containerBuilder;
+    }
+
+    /**
+     * @return CompilerPassInterface[]
+     */
+    private function createCompilerPasses(): array
+    {
+        return [
+
+            // must run before AutowireArrayParameterCompilerPass, as the autowired array cannot contain removed services
+            new RemoveSkippedRectorsCompilerPass(),
+
+            // autowire Rectors by default (mainly for tests)
+            new AutowireRectorCompilerPass(),
+            new MakeRectorsPublicCompilerPass(),
+
+            // add all merged arguments of Rector services
+            new MergeImportedRectorConfigureCallValuesCompilerPass($this->configureCallValuesCollector),
+            new AutowireArrayParameterCompilerPass(),
+        ];
+    }
+
+}
